server/requirements.txt
Django==1.8.5
argparse==1.2.1
djangorestframework==3.3.0
wsgiref==0.1.2
MySQL-python==1.2.5
django-cors-headers==1.1.0
Pillow==1.7.8
django-multiupload==0.5
flake8==2.5.0
